changes in project structure
