+++ main.py
@@ -6,6 +6,12 @@ def add_logo():
     st.markdown(
         """
         <style>
+             [data-testid="stSidebarNav"] {
+                background-image: url();
+                background-repeat: no-repeat;
+                padding-top: 10px;
+                background-position: 20px 20px;
+            }
             [data-testid="stSidebarNav"]::before {
                 content: "📑 Test2Text";
                 margin-left: 20px;
@@ -21,13 +27,13 @@ def add_logo():
 
 
 if __name__ == "__main__":
-    st.set_page_config(page_title="📑 Test2Text App", layout="wide", initial_sidebar_state="auto")
+    st.set_page_config(page_title="Test2Text App", layout="wide", initial_sidebar_state="auto")
     add_logo()
 
-    annotations = st.Page("pages/upload/annotations.py", title="Annotations", icon=":material/database_upload:")
-    requirements = st.Page("pages/upload/requirements.py", title="Requirements", icon=":material/database_upload:")
-    report = st.Page("pages/report.py", title="Report", icon=":material/publish:")
-    visualization = st.Page("services/visualisation/visualize_vectors.py", title="Visualize vectors", icon=":material/dataset:")
+    annotations = st.Page("test2text/pages/upload/annotations.py", title="Annotations", icon=":material/database_upload:")
+    requirements = st.Page("test2text/pages/upload/requirements.py", title="Requirements", icon=":material/database_upload:")
+    report = st.Page("test2text/pages/report.py", title="Report", icon=":material/publish:")
+    visualization = st.Page("test2text/services/visualisation/visualize_vectors.py", title="Visualize vectors", icon=":material/dataset:")
     pages = {
         "Upload": [annotations, requirements],
         "Inspect": [report, visualization],
