--- pages/upload/__init__.py
@@ -1,4 +1,0 @@
-from pages.upload import annotations
-from pages.upload import requirements
-
-
