+++ test2text/services/db/client.py
@@ -2,7 +2,7 @@ import sqlite3
 import sqlite_vec
 import logging
 
-from services.utils.semver import Semver
+from test2text.services.utils.semver import Semver
 from .tables import (
     RequirementsTable,
     AnnotationsTable,
