+++ test2text/services/db/tables/abstract_table.py
