+++ test2text/services/loaders/index_requirements.py
@@ -3,8 +3,8 @@ import io
 
 import streamlit as st
 
-from services.db import DbClient
-from services.embeddings.embed import embed_requirements_batch
+from test2text.services.db import DbClient
+from test2text.services.embeddings.embed import embed_requirements_batch
 
 BATCH_SIZE = 100
