+++ test2text/services/utils/sqlite_vec.py
